Rewritten logic
Rewritten logic.

In some cases the computer overwrote the player input and in some cases it was making illogical moves.

This is now corrected insofar as the human player starts the game.

Minor changes to the styling were made.

The Console was completley rewritten and now updates correctly using the useContext hook. The dispatches are made from the Gameboard component.

the reducer cases were also rewritten.